requirements.txt
django
djangorestframework
dotenv
psycopg2
djangorestframework-simplejwt
